Correction test

+++ src/TestProjet.java
@@ -40,7 +40,7 @@ public class TestProjet {
 
 			Coord coord = new Coord(1, 1);
 
-			assertTrue(coord.distance(new Coord(3, 3)) == (Math.sqrt(8)));
+			assertTrue(coord.distance(new Coord(3, 3)) == (int)(Math.sqrt(8)));
 			assertTrue(coord.distance(new Coord(1, 5)) == (Math.sqrt(16)));
 
 		}
